Hotfix for _NET_CURRENT_DESKTOP not actually being defined until a workspace has been switched.
This fix just assumes an undefined value is assumed to be on the first, as you would expect starting a session after turning on your computer.

However, I am unaware if gnome offers starting on non-first workspaces, or if you resume after a hibernation on a different workspace while the value is reset.